docs: add Contributor Covenant Code of Conduct for community standards (#11)
# reposearch PR

## Description

This establishes clear expectations for behavior, fostering a safe and
inclusive environment for all contributors, which is essential for
community growth and collaboration.

Also seeing if this solves the issue outlined here. I seem to have a
problem when updating `.github` workflows in the PR that expects a
change to them. https://github.com/cli/cli/issues/9514

## Type of Change

<!-- links -->
[1]: ../CONTRIBUTING.md
[2]: ../CONTRIBUTING.md#pr-title