docs: enhance code comments for better clarity and maintainability (#16)
# reposearch PR

## Description

Improving comments throughout the codebase aids developers in
understanding the purpose and functionality of various components. This
change reduces the onboarding time for new contributors and minimizes
the risk of misinterpretation, ultimately leading to a more robust and
maintainable codebase.

This should also trigger enhancements to the new release-by-component CI
flow.